+++ src/main/java/com/google/googleidentity/servlet/ClientServlet.java
@@ -14,22 +14,18 @@
     limitations under the License.
 */
 
-package com.google.googleidentity.resource;
+package com.google.googleidentity.servlet;
 
 import com.google.common.base.Preconditions;
 import com.google.googleidentity.oauth2.client.ClientDetails;
-import com.google.googleidentity.oauth2.token.OAuth2TokenService;
 import com.google.googleidentity.oauth2.util.OAuth2Utils;
-import com.google.googleidentity.security.UserSession;
-import com.google.googleidentity.user.UserDetails;
-import com.google.inject.Inject;
+import com.google.googleidentity.user.UserSession;
 import com.google.inject.Singleton;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import freemarker.template.Version;
 
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -43,7 +39,7 @@ import java.util.logging.Logger;
 
 /**
  * Demo UserServlet Read UserDetails.User Object {@link com.google.googleidentity.user.UserDetails}
- * stored in in the session through class {@link com.google.googleidentity.security.UserSession} and
+ * stored in in the session through class {@link UserSession} and
  * display the username.
  */
 @Singleton
