+++ src/main/java/com/google/googleidentity/servlet/LoginServlet.java
@@ -14,30 +14,32 @@
     limitations under the License.
 */
 
-package com.google.googleidentity.security;
+package com.google.googleidentity.servlet;
 
 import com.google.inject.Singleton;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import freemarker.template.Version;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
+/** Demo Login Servlet Just Bind the login Servlet with a freemarker template. */
 @Singleton
-public class ClientRegisterServlet extends HttpServlet {
+public final class LoginServlet extends HttpServlet {
 
-  private static final long serialVersionUID = 22L;
+  private static final long serialVersionUID = 1L;
 
-  private static final Logger log = Logger.getLogger("ClientRegisterServlet");
+  private static final Logger log = Logger.getLogger("LoginServlet");
 
   private Configuration configuration;
 
@@ -53,9 +55,9 @@ public class ClientRegisterServlet extends HttpServlet {
       throws ServletException, IOException {
 
     try {
-      displayPage(response);
+      displayLoginPage(response);
     } catch (TemplateException e) {
-      log.log(Level.INFO, "Error when display page", e);
+      log.log(Level.INFO, "Error when display login page", e);
     }
   }
 
@@ -63,16 +65,16 @@ public class ClientRegisterServlet extends HttpServlet {
       throws ServletException, IOException {
 
     try {
-      displayPage(response);
+      displayLoginPage(response);
     } catch (TemplateException e) {
-      log.log(Level.INFO, "Error when display page", e);
+      log.log(Level.INFO, "Error when display login page", e);
     }
   }
 
-  private void displayPage(HttpServletResponse response)
+  private void displayLoginPage(HttpServletResponse response)
       throws ServletException, IOException, TemplateException {
 
-    Template template = configuration.getTemplate("ClientRegister.ftl");
+    Template template = configuration.getTemplate("Login.ftl");
     Map<String, Object> information = new HashMap<>();
     response.setCharacterEncoding("utf-8");
     PrintWriter printWriter = response.getWriter();
