+++ src/main/java/com/google/googleidentity/servlet/UserServlet.java
@@ -14,48 +14,57 @@
     limitations under the License.
 */
 
-package com.google.googleidentity.security;
+package com.google.googleidentity.servlet;
 
+import com.google.common.base.Preconditions;
+import com.google.googleidentity.oauth2.util.OAuth2Utils;
+import com.google.googleidentity.user.UserDetails;
+import com.google.googleidentity.user.UserSession;
 import com.google.inject.Singleton;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import freemarker.template.Version;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
+/**
+ * Demo UserServlet Read UserDetails.User Object {@link com.google.googleidentity.user.UserDetails}
+ * stored in in the session through class {@link UserSession} and
+ * display the username.
+ */
 @Singleton
-public class RegisterServlet extends HttpServlet {
+public final class UserServlet extends HttpServlet {
 
-  private static final long serialVersionUID = 15L;
+  private static final long serialVersionUID = 1L;
 
-  private static final Logger log = Logger.getLogger("RegisterServlet");
-
+  private static final Logger log = Logger.getLogger("UserServlet");
   private Configuration configuration;
 
+
   public void init() throws ServletException {
+
     Version version = new Version("2.3.30");
-
     configuration = new Configuration(version);
-
     configuration.setServletContextForTemplateLoading(getServletContext(), "template");
   }
 
-  protected void doGet(HttpServletRequest request, HttpServletResponse response)
+  public void doGet(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException {
 
     try {
-      displayPage(response);
+      displayMainPage(request, response);
     } catch (TemplateException e) {
-      log.log(Level.INFO, "Error when display register page", e);
+      log.log(Level.INFO, "MainPage Error!", e);
     }
   }
 
@@ -63,20 +72,33 @@ public class RegisterServlet extends HttpServlet {
       throws ServletException, IOException {
 
     try {
-      displayPage(response);
+      displayMainPage(request, response);
     } catch (TemplateException e) {
-      log.log(Level.INFO, "Error when display register page", e);
+      log.log(Level.INFO, "MainPage Error!", e);
     }
   }
 
-  private void displayPage(HttpServletResponse response)
+  private void displayMainPage(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException, TemplateException {
 
-    Template template = configuration.getTemplate("Register.ftl");
+    UserSession userSession = OAuth2Utils.getUserSession(request);
+
+    Preconditions.checkArgument(
+        userSession.getUser().isPresent(), "User should have been logged in already");
+
+    UserDetails user = userSession.getUser().get();
+
     Map<String, Object> information = new HashMap<>();
+
+    information.put("username", user.getUsername());
+
+    Template template = configuration.getTemplate("MainPage.ftl");
+
     response.setCharacterEncoding("utf-8");
     PrintWriter printWriter = response.getWriter();
+
     template.process(information, printWriter);
+
     printWriter.flush();
   }
 }
