+++ src/main/webapp/WEB-INF/appengine-web.xml
@@ -9,7 +9,7 @@
         <env-var name="DB_USER" value="root" />
         <env-var name="DB_PASS" value="123456789" />
         <env-var name="DB_NAME" value="oauth2demo" />
-        <env-var name="USE_CLOUD_SQL" value="true" />
+        <env-var name="USE_CLOUD_SQL" value="false" />
         <env-var name="CLEAR_TABLES" value="true" />
     </env-variables>
     <system-properties>
