+++ src/test/java/com/google/googleidentity/user/UserSessionTest.java
@@ -14,11 +14,12 @@
     limitations under the License.
 */
 
-package com.google.googleidentity.security;
+package com.google.googleidentity.user;
 
 import com.google.common.base.Charsets;
 import com.google.common.hash.Hashing;
 import com.google.googleidentity.user.UserDetails;
+import com.google.googleidentity.user.UserSession;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
